requirements.txt
fastapi==0.115.6
Pillow==11.0.0
pydantic==2.10.4
pytest==7.4.4
pytest
Requests==2.32.3
ultralytics==8.3.55
